requirements.txt
httpx>=0.24
pydantic>=1.10
jinja2>=3.1
pyyaml>=6.0
pytest>=7.0
fastapi>=0.95
uvicorn>=0.22
